cloudshell_shell_core cleanup from inject
