--- cloudshell/configuration/cloudshell_shell_core_binding_keys.py
@@ -1,13 +1,0 @@
-"""Binding keys"""
-
-"""Config binding key"""
-CONFIG = 'config'
-
-"""Context binding key"""
-CONTEXT = 'context'
-
-"""Logger binding key"""
-LOGGER = 'logger'
-
-"""Cloudshell API binding key"""
-API = 'api'
