+++ requirements.txt
@@ -1,2 +1,1 @@
 cloudshell-core>=2.1,<2.2
-inject==3.3.1
